refactor: extract battle_loop.py from __init__.py
Move the core battle loop (~305 lines) into battle_loop.py.
Replace scattered global variables with a BattleState dataclass.
on_review_card() and _get_cards_per_round() now live in their own
module, importing singletons directly.
